Fix MSVC build errors (C2512, C2668, C4244)
Fix 1: StateRef default constructor (C2512)
- Added default constructor to StateRef template class
- Member variables can now be declared without immediate initialization
- Added null checks in get() and set() for safety
- Allows StateRef members to be initialized later in onMount()

Fix 2: useEffect overload ambiguity (C2668)
- Removed ambiguous lambda return syntax: return []() {};
- Use simple void-returning useEffect instead of cleanup version
- Compiler can now clearly distinguish between overloads
- PlayerComponent.h and EnemyComponent.h updated

Fix 3: Implicit int to float conversion warning (C4244)
- Added static_cast<float> to rand() expression
- Changed literal 50 to 50.0f for consistency
- Eliminates "possible loss of data" warning

These fixes resolve all MSVC C++ compilation errors while maintaining
the React-inspired architecture's functionality.

+++ src/main.cpp
@@ -43,7 +43,7 @@ void handleEvents(sf::Event& event, Store<GameState>& store) {
                     static int enemyIdCounter = 0;
                     EnemyData enemy;
                     enemy.id = enemyIdCounter++;
-                    enemy.position = sf::Vector2f(rand() % 700 + 50, 50);
+                    enemy.position = sf::Vector2f(static_cast<float>(rand() % 700 + 50), 50.0f);
                     enemy.health = 50;
                     enemy.speed = 2.0f;
